munmap errno

+++ kernel/proc/proc_manager.hh
@@ -84,9 +84,9 @@ namespace proc
         int fstat(int fd, fs::Kstat *buf);
         int chdir(eastl::string &path);
         int getcwd(char *out_buf);
-        int validate_mmap_params(void *addr, int length, int prot, int flags, int fd, int offset);
-        void *mmap(void *addr, int length, int prot, int flags, int fd, int offset, int *errno);
-        int munmap(void *addr, int length);
+        int validate_mmap_params(void *addr, size_t length, int prot, int flags, int fd, int offset);
+        void *mmap(void *addr, size_t length, int prot, int flags, int fd, int offset, int *errno);
+        int munmap(void *addr, size_t length);
         int mremap(void *old_address, size_t old_size, size_t new_size, int flags, void *new_address, void **result_addr);
         int unlink(int fd, eastl::string path, int flags);
         int pipe(int *fd, int);
